requirements.txt
PyQt5>=5.15
Pillow>=8.0
gitpython>=3.1
watchdog>=2.0